requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0
gradio>=4.0.0
scikit-image>=0.19.0
scipy>=1.7.0
matplotlib>=3.5.0
tqdm>=4.62.0
albumentations>=1.3.0
SimpleITK>=2.2.0
pydicom>=2.3.0
